pyproject.toml
[tool.poetry]
name = "paperwhisperer"
version = "0.1.0"
description = "The AI to make sense of the dense"
authors = ["JordieB <jordie.r.belle@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^0.27.8"
streamlit = "^1.23.1"
streamlit-chat = "^0.0.2.2"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
